nixos/boot: distribute systemd-time-wait-sync.service
See:

    https://github.com/systemd/systemd/issues/5097
    https://github.com/systemd/systemd/pull/8494

The basic gist here is that by default, time-sync.target only is
activated upon the *start* of systemd-timesyncd.service, which may
restore a slightly wrong system clock that needs to be adjusted when
initial measurements complete.

The systemd-time-wait-sync service, instead, waits until the kernel
reports a remote synchronized system time and then releases and allows
time-sync.target to trigger. This means that services which require
synchronized clocks can reliably wait on time-sync.target to trigger
before proceeding (if the user is using systemd-timesyncd).

Signed-off-by: Austin Seipp <aseipp@pobox.com>